rename reward classifier
