+++ tests/policies/hilserl/test_modeling_classifier.py
@@ -1,9 +1,7 @@
 import torch
 
-from lerobot.common.policies.hilserl.classifier.modeling_classifier import (
-    ClassifierConfig,
-    ClassifierOutput,
-)
+from lerobot.common.policies.reward_model.configuration_classifier import RewardClassifierConfig
+from lerobot.common.policies.reward_model.modeling_classifier import ClassifierOutput
 from lerobot.configs.types import FeatureType, NormalizationMode, PolicyFeature
 from tests.utils import require_package
 
@@ -23,11 +21,9 @@ def test_classifier_output():
 
 @require_package("transformers")
 def test_binary_classifier_with_default_params():
-    from lerobot.common.policies.hilserl.classifier.modeling_classifier import (
-        Classifier,
-    )
+    from lerobot.common.policies.reward_model.modeling_classifier import Classifier
 
-    config = ClassifierConfig()
+    config = RewardClassifierConfig()
     config.input_features = {
         "observation.image": PolicyFeature(type=FeatureType.VISUAL, shape=(3, 224, 224)),
     }
@@ -66,12 +62,10 @@ def test_binary_classifier_with_default_params():
 
 @require_package("transformers")
 def test_multiclass_classifier():
-    from lerobot.common.policies.hilserl.classifier.modeling_classifier import (
-        Classifier,
-    )
+    from lerobot.common.policies.reward_model.modeling_classifier import Classifier
 
     num_classes = 5
-    config = ClassifierConfig()
+    config = RewardClassifierConfig()
     config.input_features = {
         "observation.image": PolicyFeature(type=FeatureType.VISUAL, shape=(3, 224, 224)),
     }
@@ -107,11 +101,9 @@ def test_multiclass_classifier():
 
 @require_package("transformers")
 def test_default_device():
-    from lerobot.common.policies.hilserl.classifier.modeling_classifier import (
-        Classifier,
-    )
+    from lerobot.common.policies.reward_model.modeling_classifier import Classifier
 
-    config = ClassifierConfig()
+    config = RewardClassifierConfig()
     assert config.device == "cpu"
 
     classifier = Classifier(config)
@@ -121,11 +113,9 @@ def test_default_device():
 
 @require_package("transformers")
 def test_explicit_device_setup():
-    from lerobot.common.policies.hilserl.classifier.modeling_classifier import (
-        Classifier,
-    )
+    from lerobot.common.policies.reward_model.modeling_classifier import Classifier
 
-    config = ClassifierConfig(device="cpu")
+    config = RewardClassifierConfig(device="cpu")
     assert config.device == "cpu"
 
     classifier = Classifier(config)
